% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_example.R
\name{create_example}
\alias{create_example}
\title{Create example data}
\usage{
create_example()
}
\value{
This function returns list with example data. Each list entry has filename and software information as well as a corresponding data set.
}
\description{
Example data for ProteomeDiscoverer, Spectronaut, DIA-NN and MaxQuant.
}
\details{
Example data is generated for each software for testing functions of mpwR. Each column is created in a randomized fashion. Connections between columns are not necessarily valid. E.g. column of Precursor Charges might not reflect charges of Precuror.IDs column.
}
\examples{
data <- create_example()
}
\author{
Oliver Kardell
}
