% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_experimental_design.R
\name{write_experimental_design}
\alias{write_experimental_design}
\title{Create template for experimental design}
\usage{
write_experimental_design(path)
}
\arguments{
\item{path}{Path to folder where exp_design file is generated.}
}
\value{
This function returns a csv-file with the following columns:
\itemize{
\item analysis_name - name of your analysis.
\item software - name of used software: DIA-NN, MaxQuant, PD, Spectronaut, Generic.
\item path_to_folder - path to analysis folder.
}
}
\description{
Generation of a exp_design.csv file for using the import option with load_experimental_design.
}
\details{
The generated exp_design.csv file can be used as starting point for importing with the load_experimental_design option for mpwR. Example entries are provided. The template file - exp_design.csv - is generated under the specified path.
}
\examples{
\dontrun{
write_experimental_design(path = "DIRECTORY_WHERE_FILE_IS_GENERATED")
}
}
\author{
Oliver Kardell
}
