% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_generic_template.R
\name{write_generic_template}
\alias{write_generic_template}
\title{Create generic template}
\usage{
write_generic_template(path_filename)
}
\arguments{
\item{path_filename}{Path to folder where template is generated and user-defined filename}
}
\value{
This function returns a csv-file.
}
\description{
Generation of a template.csv file for generic input data.
}
\details{
The generated template.csv file can be used to create a software-independent input file for mpwR. Example entries are provided. The template file - filename_Report.csv - is generated. The appendix "_Report" is required for importing with mpwR.
}
\examples{
\dontrun{
write_generic_template(path = "DIRECTORY_WHERE_FILE_IS_GENERATED/filename")
}
}
\author{
Oliver Kardell
}
