% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mpwR.R
\name{prepare_mpwR}
\alias{prepare_mpwR}
\title{Load and Prepare the input data}
\usage{
prepare_mpwR(path, diann_addon_pg_qval = 0.01, diann_addon_prec_qval = 0.01)
}
\arguments{
\item{path}{Path to folder where the input data is stored - only input data. No subfolders or other files. Analysis name as prefix + for MaxQuant: _evidence, _peptides, _proteinGroups; for PD - R-friendly headers enabled: _PSMs, _Proteins, _PeptideGroups, _ProteinGroups; for DIA-NN, Spectronaut and Generic: _Report}

\item{diann_addon_pg_qval}{Numeric between 0 and 1. Applied only to DIA-NN data: \code{diann_addon_pg_qval} <= PG.Q.Value.}

\item{diann_addon_prec_qval}{Numeric between 0 and 1. Applied only to DIA-NN data: \code{diann_addon_prec_qval} <= Q.Value.}
}
\value{
A list - each list entry has filename and software info as well as stored data.
}
\description{
Input data will be imported, renamed and default filtering will be applied
}
\details{
Function for easily importing the default software outputs and preparing for downstream analysis with mpwR within one folder. As default for MaxQuant "Reverse", "Potential contaminants" and "Only identified by site" are filtered out. As default for PD only "High" confidence identifications are included and for Found in Sample column(s) also only "High" identifications. Contaminants are filtered out. As default for Spectronaut only EG.Identified equals TRUE are included.
}
\examples{
\dontrun{
prepare_mpwR(path = "DIRECTORY_TO_FILES")
}
}
\author{
Oliver Kardell
}
